# contract/openapi.yaml
openapi: 3.0.3
info:
  title: AIDetector API
  version: 1.0.0
  description: V1.0 active contract baseline for auth, detection, history, quota, admin, report, key, and team APIs.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8000
    description: Local
...
paths:
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: User login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Register user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Register success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: Email or name conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Get current user
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/me/profile:
    patch:
      tags: [auth]
      summary: Update current user profile
      operationId: updateMyProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - detail
      properties:
        code:
          oneOf:
            - type: integer
            - type: string
        message:
          type: string
        detail:
          nullable: true
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true
            - type: string
    TokenResponse:
      type: object
      required:
        - accessToken
        - tokenType
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: bearer
    LoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: Email or account name
        password:
          type: string
          format: password
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        password:
          type: string
          format: password
          minLength: 8
    UserSystemRole:
      type: string
      enum: [VISITOR, INDIVIDUAL, TEAM_ADMIN, SYS_ADMIN]
...
    UserProfile:
      type: object
      properties:
        firstName:
          type: string
          nullable: true
        surname:
          type: string
          nullable: true
        organization:
          type: string
          nullable: true
        jobRole:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
    UserProfileUpdateRequest:
      type: object
      properties:
        firstName:
          type: string
          nullable: true
        surname:
          type: string
          nullable: true
        organization:
          type: string
          nullable: true
        jobRole:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
    UserResponse:
      type: object
      required:
        - id
        - email
        - name
        - systemRole
        - isActive
        - planTier
        - creditsRemaining
        - createdAt
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
        systemRole:
          $ref: '#/components/schemas/UserSystemRole'
        isActive:
          type: boolean
        planTier:
          type: string
        creditsRemaining:
          type: integer
        createdAt:
          type: string
          format: date-time
        profile:
          $ref: '#/components/schemas/UserProfile'
...